# Fix `ShowFPS` of `D2DControl` and some performance improvement