--- a/src/D2DWinForm/D2DControl.cs
+++ b/src/D2DWinForm/D2DControl.cs
@@ -49,7 +49,7 @@ namespace unvell.D2DLib.WinForm
 		private int currentFps = 0;
 		private int lastFps = 0;
 		public bool ShowFPS { get; set; }
-		private DateTime lastFpsUpdate = DateTime.Now;
+		private DateTime lastFpsUpdate = DateTime.UtcNow;
 
 		protected override void CreateHandle()
 		{
@@ -101,10 +101,11 @@ namespace unvell.D2DLib.WinForm
 			
 			if (ShowFPS)
 			{
-				if (this.lastFpsUpdate.Second != DateTime.Now.Second)
+				if (this.lastFpsUpdate.Second != DateTime.UtcNow.Second)
 				{
 					this.lastFps = this.currentFps;
 					this.currentFps = 0;
+					this.lastFpsUpdate = DateTime.UtcNow;
 				}
 				else
 				{